feat(prussia_rt): Add Address Load exception handler
